Kivy mock moved to src/mock dir and changed symlink
