+++ keys.dat
@@ -1,0 +1,6 @@
+[BM-2cUgQGcTLWAkC6dNsv2Bc8XB3Y1GEesVLV]
+label = sad
+enabled = true
+privsigningkey = 5KWXwYq1oJMzghUSJaJoWPn8VdeBbhDN8zFot1cBd6ezKKReqBd
+privencryptionkey = 5JaeFJs8iPcQT3N8676r3gHKvJ5mTWXy1VLhGCEDqRs4vpvpxV8
+
