+++ src/bitmessagekivy/main.kv
@@ -216,8 +216,52 @@ MDNavigationLayout:
         id: scr_mngr
         size_hint_y: None
         height: root.height - toolbar.height
+        # Inbox:
+        #     id:sc1
+#         # Page:
+#         #     id:sc2
+        # Create:
+        #     id:sc3
+        # Sent:
+        #     id:sc4
+        # Trash:
+        #     id:sc5
+        Login:
+            id:sc6
+        Random:
+            id:sc7
+        # Spam:
+        #     id:sc8
+        # Setting:
+        #     id:sc9
+        MyAddress:
+            id:sc10
+        # AddressBook:
+        #     id:sc11
+        # Payment:
+        #     id:sc12
         NetworkStat:
             id:sc13
+        # MailDetail:
+        #     id:sc14
+        ShowQRCode:
+            id:sc15
+        Draft:
+            id:sc16
+        Allmails:
+            id:sc17
+        # Credits:
+        #     id:sc18
+        # Starred:
+        #     id:sc19
+        # Archieve:
+        #     id:sc20
+        # ChatRoom:
+        #     id:sc21
+        # ChatList:
+        #     id:sc22
+        ScanScreen:
+            id:sc23
 
     MDNavigationDrawer:
         id: nav_drawer
