+++ src/mock/baseclass/draft.py
@@ -22,7 +22,7 @@ from baseclass.common import (
 from kivymd.uix.label import MDLabel
 
 from pybitmessage import state
-from pybitmessage.bitmessagekivy import kivy_state
+from pybitmessage import kivy_state
 
 
 class Draft(Screen):
