+++ src/mock/baseclass/sent.py
@@ -20,7 +20,7 @@ from baseclass.common import (
 from kivymd.uix.label import MDLabel
 
 from pybitmessage import state
-from pybitmessage.bitmessagekivy import kivy_state
+from pybitmessage import kivy_state
 
 
 class Sent(Screen):
