+++ src/mock/kv
@@ -1,0 +1,1 @@
+../bitmessagekivy/kv
