+++ src/mock/main.kv
@@ -1,0 +1,1 @@
+../bitmessagekivy/main.kv
