--- src/mock/mockbitmessagekivy/kv
@@ -1,1 +1,0 @@
-../../../../../PyBitmessage/src/bitmessagekivy/kv
