--- src/mock/mockbitmessagekivy/network/threads.py
@@ -1,1 +1,0 @@
-../../../../network/threads.py
