+++ src/mock/pybitmessage/addresses.py
@@ -1,0 +1,1 @@
+../../addresses.py
