+++ src/mock/pybitmessage/bmconfigparser.py
@@ -1,0 +1,1 @@
+../../bmconfigparser.py
