+++ src/mock/pybitmessage/get_platform.py
