+++ src/mock/pybitmessage/inventory.py
