+++ src/mock/pybitmessage/network/threads.py
@@ -1,0 +1,1 @@
+../../../network/threads.py
