+++ src/mock/pybitmessage/singleton.py
@@ -1,0 +1,1 @@
+../../singleton.py
