+++ src/mock/pybitmessage/singleton1.py
